requirements.txt
flet==0.27.6
fpdf2==2.8.2
